+++ packages/cloud-cognitive/src/components/AddSelect/AddSelect.js
@@ -1,5 +1,5 @@
 //
-// Copyright IBM Corp. 2021
+// Copyright IBM Corp. 2022
 //
 // This source code is licensed under the Apache-2.0 license found in the
 // LICENSE file in the root directory of this source tree.
@@ -8,26 +8,18 @@
 import React, { forwardRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import {
-  Checkbox,
-  RadioButton,
-  StructuredListRow,
-  StructuredListWrapper,
-  StructuredListBody,
-  StructuredListCell,
-  TextInput,
-  Tag,
-} from 'carbon-components-react';
-import { ChevronRight16 } from '@carbon/icons-react';
+import { TextInput, Tag } from 'carbon-components-react';
 import { Tearsheet, TearsheetNarrow } from '../../components/Tearsheet';
 import { NoDataEmptyState } from '../../components/EmptyStates/NoDataEmptyState';
 import { pkg } from '../../settings';
+import { AddSelectSidebar } from './AddSelectSidebar';
+import { AddSelectBreadcrumbs } from './AddSelectBreadcrumbs';
+import { AddSelectList } from './AddSelectList';
 const componentName = 'AddSelect';
 
 export let AddSelect = forwardRef(
   (
     {
-      actions,
       className,
       description,
       influencerTitle,
@@ -39,23 +31,23 @@ export let AddSelect = forwardRef(
       noResultsTitle,
       noSelectionDescription,
       noSelectionTitle,
+      onClose,
+      onCloseButtonText,
       onSearchFilter,
+      onSubmit,
+      onSubmitButtonText,
       open,
+      removeIconDescription,
+      textInputLabel,
       title,
       ...rest
     },
     ref
   ) => {
     const blockClass = `${pkg.prefix}--add-select`;
-    const commonTearsheetProps = {
-      open,
-      title,
-      actions,
-      description,
-      closeIconDescription: 'temp description',
-    };
 
     // hooks
+    const [path, setPath] = useState([]);
     const [singleSelection, setSingleSelection] = useState('');
     const [multiSelection, setMultiSelection] = useState([]);
     const [searchTerm, setSearchTerm] = useState('');
@@ -65,27 +57,21 @@ export let AddSelect = forwardRef(
       setSearchTerm(e.target.value);
     };
 
-    const handleSingleSelection = (value) => {
-      setSingleSelection(value);
-    };
-
-    const handleMultiSelection = (value, checked) => {
-      if (checked) {
-        const newValues = [...multiSelection, value];
-        setMultiSelection(newValues);
-      } else {
-        const newValues = multiSelection.filter((v) => v !== value);
-        setMultiSelection(newValues);
-      }
-    };
-
-    const onNavigateItem = () => {
-      // TODO figure out navigation
-    };
-
     // item filtering
-    const getFilteredItems = () =>
-      items.filter((item) => {
+    const getFilteredItems = () => {
+      const hasPath = path.length > 0;
+      /**
+       * how to traverse the levels of items-
+       * the path represents the ids of each level / item / breadcrumb
+       * using this path we can drill down into the items until we get to the last one the user selected
+       */
+      const itemsToFilter = hasPath
+        ? path.reduce(
+            (prev, cur) => prev.find((item) => item.id === cur.id).children,
+            items
+          )
+        : items;
+      const results = itemsToFilter.filter((item) => {
         if (!searchTerm) {
           return item;
         }
@@ -94,33 +80,12 @@ export let AddSelect = forwardRef(
           return onSearchFilter(item, searchTerm);
         }
         // otherwise use the default label filter
-        return item.label.toLowerCase().includes(searchTerm);
+        return item.title.toLowerCase().includes(searchTerm);
       });
+      return results;
+    };
 
     const filteredItems = getFilteredItems();
-
-    // sidebar
-    const influencer = (
-      <div className={`${blockClass}__influencer`}>
-        <div className={`${blockClass}__influencer-header`}>
-          <p className={`${blockClass}__influencer-title`}>{influencerTitle}</p>
-          <Tag type="gray" size="sm">
-            {multiSelection.length}
-          </Tag>
-        </div>
-        <div className={`${blockClass}__influencer-body`}>
-          {multiSelection.length > 0 ? (
-            <p>content</p>
-          ) : (
-            <NoDataEmptyState
-              subtitle={noSelectionDescription}
-              title={noSelectionTitle}
-              size="sm"
-            />
-          )}
-        </div>
-      </div>
-    );
 
     // main content
     const body = (
@@ -128,60 +93,41 @@ export let AddSelect = forwardRef(
         <div className={`${blockClass}__header`}>
           <TextInput
             id="temp-id"
-            labelText="temp label"
+            labelText={textInputLabel}
             placeholder={inputPlaceholder}
             value={searchTerm}
             onChange={handleSearch}
           />
           <div className={`${blockClass}__items-label-container`}>
-            <p className={`${blockClass}__items-label`}>{itemsLabel}</p>
-            <Tag type="gray" size="sm">
+            {path.length ? (
+              <AddSelectBreadcrumbs
+                itemsLabel={itemsLabel}
+                path={path}
+                setPath={setPath}
+              />
+            ) : (
+              <p className={`${blockClass}__items-label`}>{itemsLabel}</p>
+            )}
+            <Tag
+              type="gray"
+              size="sm"
+              className={`${blockClass}__items-label-tag`}
+            >
               {filteredItems.length}
             </Tag>
           </div>
         </div>
         {filteredItems.length > 0 ? (
-          <div className={`${blockClass}__selections-wrapper`}>
-            <StructuredListWrapper
-              selection
-              className={`${blockClass}__selections`}
-            >
-              <StructuredListBody>
-                {filteredItems.map((item) => (
-                  <StructuredListRow key={item.id}>
-                    <StructuredListCell>
-                      <div className={`${blockClass}__selections-cell-wrapper`}>
-                        {multi ? (
-                          <Checkbox
-                            className={`${blockClass}__selections-checkbox`}
-                            onChange={(value) =>
-                              handleMultiSelection(item.value, value)
-                            }
-                            labelText={item.label}
-                            id={item.id}
-                            checked={multiSelection.includes(item.value)}
-                          />
-                        ) : (
-                          <RadioButton
-                            className={`${blockClass}__selections-radio`}
-                            name="add-select-selections"
-                            id={item.id}
-                            value={item.value}
-                            labelText={item.label}
-                            onChange={handleSingleSelection}
-                            selected={item.value === singleSelection}
-                          />
-                        )}
-                        {item.children && (
-                          <ChevronRight16 onClick={onNavigateItem} />
-                        )}
-                      </div>
-                    </StructuredListCell>
-                  </StructuredListRow>
-                ))}
-              </StructuredListBody>
-            </StructuredListWrapper>
-          </div>
+          <AddSelectList
+            filteredItems={filteredItems}
+            multi={multi}
+            multiSelection={multiSelection}
+            path={path}
+            setMultiSelection={setMultiSelection}
+            setPath={setPath}
+            setSingleSelection={setSingleSelection}
+            singleSelection={singleSelection}
+          />
         ) : (
           <div className={`${blockClass}__body`}>
             <NoDataEmptyState
@@ -193,12 +139,47 @@ export let AddSelect = forwardRef(
       </>
     );
 
+    const commonTearsheetProps = {
+      open,
+      title,
+      description,
+      closeIconDescription: 'temp description',
+      actions: [
+        {
+          label: onCloseButtonText,
+          kind: 'secondary',
+          onClick: onClose,
+        },
+        {
+          label: onSubmitButtonText,
+          kind: 'primary',
+          onClick: onSubmit,
+          disabled: multi ? multiSelection.length === 0 : !singleSelection,
+        },
+      ],
+    };
+
+    const sidebarProps = {
+      influencerTitle,
+      items,
+      multiSelection,
+      noSelectionDescription,
+      noSelectionTitle,
+      removeIconDescription,
+      setMultiSelection,
+    };
+
+    const classNames = cx(className, blockClass, {
+      [`${blockClass}__single`]: !multi,
+      [`${blockClass}__multi`]: multi,
+    });
+
     return (
-      <div ref={ref} className={cx(className, blockClass)} {...rest}>
+      <div ref={ref} className={classNames} {...rest}>
         {multi ? (
           <Tearsheet
             {...commonTearsheetProps}
-            influencer={multi && influencer}
+            influencer={multi && <AddSelectSidebar {...sidebarProps} />}
             influencerPosition="right"
           >
             {body}
@@ -212,16 +193,15 @@ export let AddSelect = forwardRef(
 );
 
 AddSelect.propTypes = {
-  actions: PropTypes.array,
   className: PropTypes.string,
   description: PropTypes.string,
   influencerTitle: PropTypes.string,
   inputPlaceholder: PropTypes.string,
   items: PropTypes.arrayOf(
     PropTypes.shape({
-      id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-      label: PropTypes.string,
-      value: PropTypes.string,
+      id: PropTypes.string.isRequired,
+      title: PropTypes.string.isRequired,
+      value: PropTypes.string.isRequired,
     })
   ),
   itemsLabel: PropTypes.string,
@@ -230,8 +210,14 @@ AddSelect.propTypes = {
   noResultsTitle: PropTypes.string,
   noSelectionDescription: PropTypes.string,
   noSelectionTitle: PropTypes.string,
+  onClose: PropTypes.func,
+  onCloseButtonText: PropTypes.string,
   onSearchFilter: PropTypes.func,
+  onSubmit: PropTypes.func,
+  onSubmitButtonText: PropTypes.string,
   open: PropTypes.bool,
+  removeIconDescription: PropTypes.string,
+  textInputLabel: PropTypes.string,
   title: PropTypes.string,
 };