+++ packages/cloud-cognitive/src/components/MultiAddSelect/MultiAddSelect.stories.js
@@ -28,31 +28,24 @@ export default {
 
 const defaultProps = {
   open: true,
-  actions: [
-    {
-      label: 'Cancel',
-      kind: 'secondary',
-    },
-    {
-      label: 'Add',
-      kind: 'primary',
-    },
-  ],
   items: [
     {
       id: '1',
-      label: 'Kansas',
+      title: 'Kansas',
       value: 'kansas',
+      subtitle: 'subtitle',
     },
     {
       id: '2',
-      label: 'Texas',
+      title: 'Texas',
       value: 'texas',
+      subtitle: 'subtitle',
     },
     {
       id: '3',
-      label: 'Florida',
+      title: 'Florida',
       value: 'florida',
+      subtitle: 'subtitle',
     },
   ],
   title: 'Add business terms',
@@ -65,6 +58,10 @@ const defaultProps = {
     'Select a term to include the full set of the governance artifacts it contains in the governance scope.',
   noResultsTitle: 'No results',
   noResultsDescription: 'Try again',
+  onCloseButtonText: 'Cancel',
+  onSubmitButtonText: 'Add',
+  removeIconDescription: 'Remove',
+  textInputLabel: 'test input label',
 };
 
 const Template = (args) => {
@@ -74,5 +71,35 @@ const Template = (args) => {
 export const Default = prepareStory(Template, {
   args: {
     ...defaultProps,
+  },
+});
+
+export const WithHierarchy = prepareStory(Template, {
+  args: {
+    ...defaultProps,
+    items: [
+      ...defaultProps.items,
+      {
+        id: '4',
+        title: 'California',
+        value: 'california',
+        subtitle: 'some / other / path',
+        children: [
+          {
+            id: '5',
+            title: 'Los Angeles',
+            value: 'la',
+            children: [
+              {
+                id: '453',
+                title: 'Something',
+                value: 'something',
+                tags: 'list of tags',
+              },
+            ],
+          },
+        ],
+      },
+    ],
   },
 });